refactor(tree): made changes to file structure
