+++ docCrawler/middleware/__init__.py
@@ -1,0 +1,1 @@
+from docCrawler.middleware.middlewares import *  # noqa
