+++ docCrawler/utils/__init__.py
@@ -1,0 +1,2 @@
+from docCrawler.utils.items import *  # noqa
+from docCrawler.utils.pipelines import *  # noqa
