+++ docCrawler/utils/pipelines.py
@@ -1,0 +1,12 @@
+# Define your item pipelines here
+#
+# Don't forget to add your pipeline to the ITEM_PIPELINES setting
+# See: https://docs.scrapy.org/en/latest/topics/item-pipeline.html
+
+
+# useful for handling different item types with a single interface
+
+
+class DoccrawlerPipeline:
+    def process_item(self, item, spider):
+        return item
