--- requirements.txt
@@ -1,13 +1,0 @@
-duckduckgo-search==3.8.5
-fastapi==0.99.1
-twisted==22.10.0
-aiohttp==3.8.5
-langchain==0.0.250
-scrapy==2.9.0
-sentence-transformers==2.2.2
-jq==1.4.1
-uvicorn==0.23.2
-beautifulsoup4==4.12.2
-openai==0.27.8
-chromadb==0.4.4
-tiktoken==0.4.0
