+++ streamlit_app.py
@@ -1,21 +1,7 @@
-import multiprocessing
 import requests
-from scrapy import cmdline
 import streamlit as st
-from twisted.internet import reactor
-import json
-from scrapy import cmdline
-from docCrawler.docCrawler.spiders.spider import DocSpider
-from scrapy.crawler import CrawlerProcess
-from helpers import (
-    start_crawler,
-    get_full_text,
-    split_docs,
-    data_loader,
-    init_database,
-    search_similarity
-)
-    
+from scripts.helpers import search_similarity
+
 st.title("Chat-with-your-docs")
 
 # Sidebar
@@ -37,21 +23,20 @@ if st.button("Search"):
 # Start crawling
 if st.sidebar.button("Start Crawling"):
     st.sidebar.text("Crawling in progress...")
-        
+
     api_url = "http://localhost:8000/getLinks"
-    params = {
-        "start_url": start_url,
-        "max_depth": max_depth
-    }
+    params = {"start_url": start_url, "max_depth": max_depth}
     response = requests.get(api_url, params=params)
-    
+
     st.sidebar.text("Crawling completed!")
 # Summarize and split documents
 if st.button("Process Documents"):
     st.text("Processing documents...")
-        
+
     # Make a request to FastAPI route /getText with the specified file path
-    response_text = requests.get("http://localhost:8000/getText", params={"path": "files/output_links.json"}).json()
+    response_text = requests.get(
+        "http://localhost:8000/getText", params={"path": "files/output_links.json"}
+    ).json()
     st.text(response_text["status"])
 
     st.sidebar.text("Splitting documents...")
